+++ src/main/java/com/nhnacademy/marketgg/server/service/file/FileService.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.marketgg.server.service.image;
+package com.nhnacademy.marketgg.server.service.file;
 
 import com.nhnacademy.marketgg.server.dto.response.image.ImageResponse;
 import com.nhnacademy.marketgg.server.entity.Asset;
@@ -7,7 +7,7 @@ import java.io.IOException;
 import java.util.List;
 import org.springframework.web.multipart.MultipartFile;
 
-public interface ImageService {
+public interface FileService {
 
     List<Image> parseImages(List<MultipartFile> multipartFiles, Asset asset) throws IOException;
